updated Storage component to allow changing of storage location
When location is changed, the file at old location will be deleted
automatically.
Backup file is kept in the system directory rather than at the location
user specified so that when user changed storage location, only task
data file need to be moved.
Currently undo operation for "location" command is not allowed since
user may not remember the old location. This will also enable the user
to simply set a new storage location when user wants to sync with cloud
service but doesn't remember the  storage location that has been set
previously.