updated Storage component to allow moving completed tasks to a separate file
This is to allow user to look at the data of completed tasks
collectively if read from the text files directly and not through the
software.
With "search" command, user can find both completed or incomplete tasks.
With "delete" command after "search" or "view done", user can delete
both completed or incomplete tasks.